Generification, brings in a lot more classes but adds convenience for our own sake.
