+++ src/main/java/io/jking/tickster/core/Tickster.java
@@ -6,11 +6,11 @@ import io.jking.tickster.database.Hikari;
 import io.jking.tickster.handler.GuildHandler;
 import io.jking.tickster.handler.InteractionHandler;
 import io.jking.tickster.handler.StartHandler;
-import io.jking.tickster.interaction.slash.impl.info.AboutCommand;
-import io.jking.tickster.interaction.slash.impl.setup.SetupCommand;
-import io.jking.tickster.interaction.slash.impl.utility.PingCommand;
-import io.jking.tickster.interaction.slash.impl.utility.TestCommand;
-import io.jking.tickster.interaction.slash.object.CommandRegistry;
+import io.jking.tickster.interaction.impl.slash.impl.info.AboutCommand;
+import io.jking.tickster.interaction.impl.slash.impl.setup.SetupCommand;
+import io.jking.tickster.interaction.impl.slash.impl.utility.PingCommand;
+import io.jking.tickster.interaction.impl.slash.impl.utility.TestCommand;
+import io.jking.tickster.interaction.impl.slash.object.CommandRegistry;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.requests.GatewayIntent;
