+++ src/main/java/io/jking/tickster/interaction/InteractionImpl.java
@@ -1,0 +1,10 @@
+package io.jking.tickster.interaction;
+
+
+public interface InteractionImpl<T extends AbstractContext<?>> {
+
+    void onInteraction(T context);
+
+    String componentId();
+
+}
