--- src/main/java/io/jking/tickster/interaction/button/impl/ticket/CloseTicketButton.java
@@ -1,24 +1,0 @@
-package io.jking.tickster.interaction.button.impl.ticket;
-
-import io.jking.tickster.interaction.button.object.ButtonContext;
-import io.jking.tickster.interaction.button.object.IButton;
-import net.dv8tion.jda.api.interactions.components.Button;
-import net.dv8tion.jda.api.interactions.components.ButtonStyle;
-
-public class CloseTicketButton implements IButton {
-    @Override
-    public void onButtonPress(ButtonContext context) {
-        context.getInteraction().deferEdit().queue(hook -> context.getChannel().sendMessage("Are you sure you want to close this ticket?")
-                .setActionRow(
-                        Button.of(ButtonStyle.SUCCESS, "yes:ticket_close", "Yes"),
-                        Button.of(ButtonStyle.DANGER, "no:ticket_close", "Cancel")
-                )
-                .queue());
-    }
-
-    @Override
-    public String buttonId() {
-        return "close_ticket";
-    }
-
-}
