--- src/main/java/io/jking/tickster/interaction/button/object/ButtonRegistry.java
@@ -1,27 +1,0 @@
-package io.jking.tickster.interaction.button.object;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ButtonRegistry {
-    private final Map<String, IButton> buttonMap;
-
-    public ButtonRegistry() {
-        this.buttonMap = new HashMap<>();
-    }
-
-    public void addButton(IButton button) {
-        buttonMap.put(button.buttonId(), button);
-    }
-
-    public ButtonRegistry addButtons(IButton... buttons) {
-        for (IButton button : buttons) {
-            addButton(button);
-        }
-        return this;
-    }
-
-    public IButton getButton(String buttonId) {
-        return buttonMap.getOrDefault(buttonId, null);
-    }
-}
