--- src/main/java/io/jking/tickster/interaction/button/object/IButton.java
@@ -1,8 +1,0 @@
-package io.jking.tickster.interaction.button.object;
-
-public interface IButton {
-
-    void onButtonPress(ButtonContext context);
-
-    String buttonId();
-}
