+++ src/main/java/io/jking/tickster/interaction/impl/selection/HelpSelection.java
@@ -1,0 +1,17 @@
+package io.jking.tickster.interaction.impl.selection;
+
+import io.jking.tickster.interaction.InteractionImpl;
+import io.jking.tickster.interaction.impl.slash.object.SelectionContext;
+
+public class HelpSelection implements InteractionImpl<SelectionContext> {
+
+    @Override
+    public void onInteraction(SelectionContext context) {
+
+    }
+
+    @Override
+    public String componentId() {
+        return null;
+    }
+}
