+++ src/main/java/io/jking/tickster/interaction/impl/slash/object/type/SuccessType.java
@@ -1,0 +1,20 @@
+package io.jking.tickster.interaction.impl.slash.object.type;
+
+public enum SuccessType {
+
+    UPDATED("%s was successfully \nupdated to %s.");
+
+    private final String successType;
+
+    SuccessType(String successType) {
+        this.successType = successType;
+    }
+
+    public String getSuccessType(Object... objects) {
+        return successType.formatted(objects);
+    }
+
+    public String getName() {
+        return name().charAt(0) + name().substring(1).toLowerCase();
+    }
+}
