--- src/main/java/io/jking/tickster/interaction/selection/object/ISelection.java
@@ -1,4 +1,0 @@
-package io.jking.tickster.interaction.selection.object;
-
-public interface ISelection {
-}
