--- src/main/java/io/jking/tickster/interaction/selection/object/SelectionRegistry.java
@@ -1,4 +1,0 @@
-package io.jking.tickster.interaction.selection.object;
-
-public class SelectionRegistry {
-}
