+++ src/main/java/io/jking/tickster/interaction/type/IButton.java
@@ -1,0 +1,7 @@
+package io.jking.tickster.interaction.type;
+
+import io.jking.tickster.interaction.InteractionImpl;
+import io.jking.tickster.interaction.context.ButtonContext;
+
+public interface IButton extends InteractionImpl<ButtonContext> {
+}
