+++ src/main/java/io/jking/tickster/interaction/type/ISelection.java
@@ -1,0 +1,8 @@
+package io.jking.tickster.interaction.type;
+
+import io.jking.tickster.interaction.InteractionImpl;
+import io.jking.tickster.interaction.context.ButtonContext;
+
+public interface ISelection extends InteractionImpl<ButtonContext> {
+
+}
